+++ packages/transaction-controller/src/TransactionController.test.ts
@@ -77,6 +77,7 @@ import {
   WalletDevice,
 } from './types';
 import { addTransactionBatch } from './utils/batch';
+import { getDelegationAddress } from './utils/eip7702';
 import { addGasBuffer, estimateGas, updateGas } from './utils/gas';
 import { updateGasFees } from './utils/gas-fees';
 import { getGasFeeFlow } from './utils/gas-flow';
@@ -120,13 +121,19 @@ jest.mock('./utils/gas');
 jest.mock('./utils/gas-fees');
 jest.mock('./utils/gas-flow');
 jest.mock('./utils/layer1-gas-fee-flow');
+jest.mock('./utils/simulation');
+jest.mock('./utils/swaps');
+jest.mock('uuid');
+
 jest.mock('./helpers/ResimulateHelper', () => ({
   ...jest.requireActual('./helpers/ResimulateHelper'),
   shouldResimulate: jest.fn(),
 }));
-jest.mock('./utils/simulation');
-jest.mock('./utils/swaps');
-jest.mock('uuid');
+
+jest.mock('./utils/eip7702', () => ({
+  ...jest.requireActual('./utils/eip7702'),
+  getDelegationAddress: jest.fn(),
+}));
 
 // TODO: Replace `any` with type
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -497,6 +504,7 @@ describe('TransactionController', () => {
   );
   const addTransactionBatchMock = jest.mocked(addTransactionBatch);
   const methodDataHelperClassMock = jest.mocked(MethodDataHelper);
+  const getDelegationAddressMock = jest.mocked(getDelegationAddress);
 
   let mockEthQuery: EthQuery;
   let getNonceLockSpy: jest.Mock;
@@ -642,11 +650,12 @@ describe('TransactionController', () => {
       unrestrictedMessenger.getRestricted({
         name: 'TransactionController',
         allowedActions: [
+          'AccountsController:getSelectedAccount',
+          'AccountsController:getState',
           'ApprovalController:addRequest',
           'NetworkController:getNetworkClientById',
           'NetworkController:findNetworkClientIdByChainId',
-          'AccountsController:getSelectedAccount',
-          'AccountsController:getState',
+          'RemoteFeatureFlagController:getState',
         ],
         allowedEvents: [],
       });
@@ -660,6 +669,15 @@ describe('TransactionController', () => {
     unrestrictedMessenger.registerActionHandler(
       'AccountsController:getState',
       () => ({}) as never,
+    );
+
+    const remoteFeatureFlagControllerGetStateMock = jest.fn().mockReturnValue({
+      featureFlags: {},
+    });
+
+    unrestrictedMessenger.registerActionHandler(
+      'RemoteFeatureFlagController:getState',
+      remoteFeatureFlagControllerGetStateMock,
     );
 
     const controller = new TransactionController({
@@ -678,6 +696,12 @@ describe('TransactionController', () => {
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       {} as any,
     );
+
+    getDelegationAddressMock.mockResolvedValue(undefined);
+
+    remoteFeatureFlagControllerGetStateMock.mockReturnValue({
+      remoteFeatureFlags: {},
+    });
 
     return {
       controller,
@@ -1583,6 +1607,7 @@ describe('TransactionController', () => {
         batchId: undefined,
         chainId: expect.any(String),
         dappSuggestedGasFees: undefined,
+        delegationAddress: undefined,
         deviceConfirmedOn: undefined,
         id: expect.any(String),
         isFirstTimeInteraction: undefined,
@@ -1932,6 +1957,26 @@ describe('TransactionController', () => {
         getSavedGasFees: expect.any(Function),
         txMeta: expect.any(Object),
       });
+    });
+
+    it('adds delegration address to metadata', async () => {
+      const { controller } = setupController();
+
+      getDelegationAddressMock.mockResolvedValueOnce(ACCOUNT_MOCK);
+
+      await controller.addTransaction(
+        {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_MOCK,
+        },
+        {
+          networkClientId: NETWORK_CLIENT_ID_MOCK,
+        },
+      );
+
+      expect(controller.state.transactions[0].delegationAddress).toBe(
+        ACCOUNT_MOCK,
+      );
     });
 
     describe('updates simulation data', () => {
@@ -2385,6 +2430,57 @@ describe('TransactionController', () => {
               status: 'rejected',
             }),
             actionId: mockActionId,
+          }),
+        );
+      });
+
+      it('publishes TransactionController:transactionRejected if error is method not supported', async () => {
+        const error = {
+          code: errorCodes.rpc.methodNotSupported,
+        };
+
+        const { controller, messenger } = setupController({
+          messengerOptions: {
+            addTransactionApprovalRequest: {
+              state: 'rejected',
+              error,
+            },
+          },
+        });
+
+        const rejectedEventListener = jest.fn();
+
+        messenger.subscribe(
+          'TransactionController:transactionRejected',
+          rejectedEventListener,
+        );
+
+        const { result } = await controller.addTransaction(
+          {
+            from: ACCOUNT_MOCK,
+            to: ACCOUNT_MOCK,
+          },
+          {
+            networkClientId: NETWORK_CLIENT_ID_MOCK,
+          },
+        );
+
+        const finishedPromise = waitForTransactionFinished(messenger);
+
+        try {
+          await result;
+        } catch {
+          // Ignore user rejected error as it is expected
+        }
+        await finishedPromise;
+
+        expect(rejectedEventListener).toHaveBeenCalledTimes(1);
+        expect(rejectedEventListener).toHaveBeenCalledWith(
+          expect.objectContaining({
+            transactionMeta: expect.objectContaining({
+              error,
+              status: 'rejected',
+            }),
           }),
         );
       });
@@ -6222,13 +6318,9 @@ describe('TransactionController', () => {
       expect(controller.state.transactions[0].status).toBe(
         TransactionStatus.failed,
       );
-      expect(
-        (
-          controller.state.transactions[0] as TransactionMeta & {
-            status: TransactionStatus.failed;
-          }
-        ).error.message,
-      ).toBe('Signing aborted by user');
+      expect(controller.state.transactions[0].error?.message).toBe(
+        'Signing aborted by user',
+      );
     });
   });