+++ packages/transaction-controller/src/index.ts
@@ -45,6 +45,7 @@ export type {
   InferTransactionTypeResult,
   LegacyGasFeeEstimates,
   Log,
+  NestedTransactionMetadata,
   SavedGasFees,
   SecurityAlertResponse,
   SecurityProviderRequest,